+++ Book/src/FictionBook.java
@@ -5,7 +5,7 @@ public class FictionBook extends Book{
     public FictionBook(String category){
         this.category = category;
     }
-    public FictionBook(String bookCode,String name, String author, double price,String category){
+    public FictionBook(String bookCode,String name, String author, int price,String category){
         super(bookCode, name, author, price);
         this.category = category;
     }
